Break commands up into OS-specific packages
