+++ internal/crypto/encrypt.go
@@ -1,0 +1,42 @@
+package crypto
+
+import (
+	"bytes"
+	"crypto/aes"
+	"crypto/cipher"
+	"encoding/base64"
+
+	"github.com/Coalfire-Research/Slackor/internal/config"
+)
+
+func PKCS5Padding(ciphertext []byte, blockSize int) []byte { //Used for Crypto
+	padding := blockSize - len(ciphertext)%blockSize
+	padtext := bytes.Repeat([]byte{byte(padding)}, padding)
+	return append(ciphertext, padtext...)
+}
+
+func Encrypt(origData []byte) (string, error) { // Encrypt a string
+	block, err := aes.NewCipher(config.CipherKeyBytes)
+	if err != nil {
+		return "", err
+	}
+	blockSize := block.BlockSize()
+	origData = PKCS5Padding(origData, blockSize)
+	blockMode := cipher.NewCBCEncrypter(block, config.CipherIV)
+	crypted := make([]byte, len(origData))
+	blockMode.CryptBlocks(crypted, origData)
+	return base64.StdEncoding.EncodeToString(crypted), nil
+}
+
+func EncryptFile(origData []byte) ([]byte, error) { // Encrypt a file
+	block, err := aes.NewCipher(config.CipherKeyBytes)
+	if err != nil {
+		return nil, err
+	}
+	blockSize := block.BlockSize()
+	origData = PKCS5Padding(origData, blockSize)
+	blockMode := cipher.NewCBCEncrypter(block, config.CipherIV)
+	crypted := make([]byte, len(origData))
+	blockMode.CryptBlocks(crypted, origData)
+	return crypted, nil
+}
