+++ internal/doc.go
@@ -1,0 +1,3 @@
+// Package internal is the collection of utility packages and commands used by
+// Slackor components which are specific to its internals.
+package internal
