+++ pkg/common/doctor.go
@@ -1,0 +1,3 @@
+package common
+
+// TODO: track errors and misconfigurations that we might otherwise panic for
