+++ pkg/common/testdata/cat/a.txt
@@ -1,0 +1,1 @@
+The quick brown fox jumped over the lazy dog.
