+++ pkg/common/testdata/cat/subdir/b.txt
@@ -1,0 +1,1 @@
+Sphinx of black quartz, judge my vow.
