+++ pkg/darwin/version.go
@@ -1,0 +1,1 @@
+package darwin
