+++ pkg/linux/sysinfo.go
@@ -1,0 +1,1 @@
+package linux
