+++ pkg/linux/version.go
@@ -1,0 +1,1 @@
+package linux
