+++ pkg/windows/const.go
@@ -1,0 +1,116 @@
+package windows
+
+const (
+	// MEM_COMMIT is a Windows constant used with Windows API calls
+	MEM_COMMIT = 0x1000
+	// MEM_RESERVE is a Windows constant used with Windows API calls
+	MEM_RESERVE = 0x2000
+	// MEM_RELEASE is a Windows constant used with Windows API calls
+	MEM_RELEASE = 0x8000
+	// PAGE_EXECUTE is a Windows constant used with Windows API calls
+	PAGE_EXECUTE = 0x10
+	// PAGE_EXECUTE_READWRITE is a Windows constant used with Windows API calls
+	PAGE_EXECUTE_READWRITE = 0x40
+	// PAGE_READWRITE is a Windows constant used with Windows API calls
+	PAGE_READWRITE = 0x04
+	// PROCESS_CREATE_THREAD is a Windows constant used with Windows API calls
+	PROCESS_CREATE_THREAD = 0x0002
+	// PROCESS_VM_READ is a Windows constant used with Windows API calls
+	PROCESS_VM_READ = 0x0010
+	//PROCESS_VM_WRITE is a Windows constant used with Windows API calls
+	PROCESS_VM_WRITE = 0x0020
+	// PROCESS_VM_OPERATION is a Windows constant used with Windows API calls
+	PROCESS_VM_OPERATION = 0x0008
+	// PROCESS_QUERY_INFORMATION is a Windows constant used with Windows API calls
+	PROCESS_QUERY_INFORMATION = 0x0400
+	// TH32CS_SNAPHEAPLIST is a Windows constant used with Windows API calls
+	TH32CS_SNAPHEAPLIST = 0x00000001
+	// TH32CS_SNAPMODULE is a Windows constant used with Windows API calls
+	TH32CS_SNAPMODULE = 0x00000008
+	// TH32CS_SNAPPROCESS is a Windows constant used with Windows API calls
+	TH32CS_SNAPPROCESS = 0x00000002
+	// TH32CS_SNAPTHREAD is a Windows constant used with Windows API calls
+	TH32CS_SNAPTHREAD = 0x00000004
+	// THREAD_SET_CONTEXT is a Windows constant used with Windows API calls
+	THREAD_SET_CONTEXT   = 0x0010
+	ERROR_ALREADY_EXISTS = 183
+
+	// Virtual-Key Codes
+	vk_BACK       = 0x08
+	vk_TAB        = 0x09
+	vk_CLEAR      = 0x0C
+	vk_RETURN     = 0x0D
+	vk_SHIFT      = 0x10
+	vk_CONTROL    = 0x11
+	vk_MENU       = 0x12
+	vk_PAUSE      = 0x13
+	vk_CAPITAL    = 0x14
+	vk_ESCAPE     = 0x1B
+	vk_SPACE      = 0x20
+	vk_PRIOR      = 0x21
+	vk_NEXT       = 0x22
+	vk_END        = 0x23
+	vk_HOME       = 0x24
+	vk_LEFT       = 0x25
+	vk_UP         = 0x26
+	vk_RIGHT      = 0x27
+	vk_DOWN       = 0x28
+	vk_SELECT     = 0x29
+	vk_PRINT      = 0x2A
+	vk_EXECUTE    = 0x2B
+	vk_SNAPSHOT   = 0x2C
+	vk_INSERT     = 0x2D
+	vk_DELETE     = 0x2E
+	vk_LWIN       = 0x5B
+	vk_RWIN       = 0x5C
+	vk_APPS       = 0x5D
+	vk_SLEEP      = 0x5F
+	vk_NUMPAD0    = 0x60
+	vk_NUMPAD1    = 0x61
+	vk_NUMPAD2    = 0x62
+	vk_NUMPAD3    = 0x63
+	vk_NUMPAD4    = 0x64
+	vk_NUMPAD5    = 0x65
+	vk_NUMPAD6    = 0x66
+	vk_NUMPAD7    = 0x67
+	vk_NUMPAD8    = 0x68
+	vk_NUMPAD9    = 0x69
+	vk_MULTIPLY   = 0x6A
+	vk_ADD        = 0x6B
+	vk_SEPARATOR  = 0x6C
+	vk_SUBTRACT   = 0x6D
+	vk_DECIMAL    = 0x6E
+	vk_DIVIDE     = 0x6F
+	vk_F1         = 0x70
+	vk_F2         = 0x71
+	vk_F3         = 0x72
+	vk_F4         = 0x73
+	vk_F5         = 0x74
+	vk_F6         = 0x75
+	vk_F7         = 0x76
+	vk_F8         = 0x77
+	vk_F9         = 0x78
+	vk_F10        = 0x79
+	vk_F11        = 0x7A
+	vk_F12        = 0x7B
+	vk_NUMLOCK    = 0x90
+	vk_SCROLL     = 0x91
+	vk_LSHIFT     = 0xA0
+	vk_RSHIFT     = 0xA1
+	vk_LCONTROL   = 0xA2
+	vk_RCONTROL   = 0xA3
+	vk_LMENU      = 0xA4
+	vk_RMENU      = 0xA5
+	vk_OEM_1      = 0xBA
+	vk_OEM_PLUS   = 0xBB
+	vk_OEM_COMMA  = 0xBC
+	vk_OEM_MINUS  = 0xBD
+	vk_OEM_PERIOD = 0xBE
+	vk_OEM_2      = 0xBF
+	vk_OEM_3      = 0xC0
+	vk_OEM_4      = 0xDB
+	vk_OEM_5      = 0xDC
+	vk_OEM_6      = 0xDD
+	vk_OEM_7      = 0xDE
+	vk_OEM_8      = 0xDF
+)
